payrollserives/src/main/java/processor/PaymentProcessor.java
package processor;

import org.springframework.stereotype.Service;

@Service
public class PaymentProcessor {
    public boolean processPayment(String cardNumber, String expirationDate, String cvv, double amount) {
        // Validate card information
        if (!isValidCardNumber(cardNumber) || !isValidExpirationDate(expirationDate) || !isValidCVV(cvv)) {
            System.out.println("Invalid card information");
            return false;
        }

        // Process payment
        if (isPaymentSuccessful(cardNumber, expirationDate, cvv, amount)) {
            System.out.println("Payment processed successfully");
            return true;
        } else {
            System.out.println("Payment failed");
            return false;
        }
    }

    private boolean isValidCardNumber(String cardNumber) {
        // Add card number validation logic here
        return cardNumber.length() == 16; // Example validation: Card number must be 16 digits long
    }

    private boolean isValidExpirationDate(String expirationDate) {
        // Add expiration date validation logic here
        // Example validation: Expiration date must be in the format "MM/YY" and should be a future date
        // You may need to use a date library or parse the string to validate the date.
        return expirationDate.matches("^(0[1-9]|1[0-2])/[0-9]{2}$");
    }

    private boolean isValidCVV(String cvv) {
        // Add CVV validation logic here
        return cvv.length() == 3; // Example validation: CVV must be 3 digits long
    }

    private boolean isPaymentSuccessful(String cardNumber, String expirationDate, String cvv, double amount) {
        // Add payment processing logic here
        // This is a placeholder implementation that assumes all payments are successful
        return true;
    }
}
